Revert "refactor(kiloclaw): share morning briefing delivery constants"
This reverts commit 384cc0d4f18ac7744c3cce27f6effad603248bc1.
